Moved all root files to separate packages.
