+++ go.mod
@@ -1,11 +1,9 @@
-module jiralert
+module github.com/free/jiralert
 
 require (
-	github.com/BurntSushi/toml v0.3.1 // indirect
 	github.com/andygrunwald/go-jira v0.0.0-20181018203616-bbce4afa5493
 	github.com/beorn7/perks v0.0.0-20160804104726-4c0e84591b9a // indirect
 	github.com/fatih/structs v0.0.0-20171020064819-f5faa72e7309 // indirect
-	github.com/free/jiralert v0.0.0-20190308154351-21520699b386
 	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
 	github.com/golang/protobuf v0.0.0-20171021043952-1643683e1b54 // indirect
 	github.com/google/go-querystring v0.0.0-20170111101155-53e6ce116135 // indirect
@@ -18,8 +16,7 @@ require (
 	github.com/prometheus/procfs v0.0.0-20171017214025-a6e9df898b13 // indirect
 	github.com/trivago/tgo v1.0.1
 	golang.org/x/sync v0.0.0-20190227155943-e225da77a7e6 // indirect
-	golang.org/x/tools v0.0.0-20190308174544-00c44ba9c14f // indirect
+	golang.org/x/tools v0.0.0-20190315214010-f0bfdbff1f9c // indirect
 	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 	gopkg.in/yaml.v2 v2.0.0-20170812160011-eb3733d160e7
-	honnef.co/go/tools v0.0.0-20190215041234-466a0476246c // indirect
 )
